Fix typos and import errors in settings.py and Categorie.py

+++ plataforma/models.py
@@ -14,7 +14,7 @@ class Resource(models.Model):
 class Product(models.Model):
     title = models.CharField(max_length=200)
     value = models.DecimalField(max_digits=10, decimal_places=2)
-    stars = models.FloatField()
+    stars = models.IntegerField()
     image = models.ImageField(upload_to='products/images/', null=True, blank=True)
     
     def __str__(self):
